+++ uc-common/infrastructure/az-data-explorer/upload.test-results.sh
@@ -14,18 +14,18 @@ stateDir="$scriptDir/state"
 stateFile=$(assertFile "$stateDir/az.adx.state.json") || exit
 echo;
 echo "##############################################################################################################"
-echo "# Script: $scriptName"
+echo "# Script: $scriptName $*"
 ##############################################################################################################################
 # Args
 runDir=$1
-if [[ ! -d "$runDir" ]]; then echo "run directory not found: '$runDir'"; exit 1; fi
+if [[ ! -d "$runDir" ]]; then echo " >>> ERROR: run directory not found: '$runDir'"; exit 1; fi
 metaFile="$runDir/run.meta.json"
-if [[ ! -f "$metaFile" ]]; then echo "meta file not found: '$metaFile'"; exit 1; fi
+if [[ ! -f "$metaFile" ]]; then echo " >>> ERROR: meta file not found: '$metaFile'"; exit 1; fi
 
 ##############################################################################################################################
 # State vars
-stateJson=$(cat $stateFile | jq)
-metaJson=$(cat $metaFile | jq)
+stateJson=$(cat $stateFile | jq -r . ) || exit
+metaJson=$(cat $metaFile | jq -r . ) || exit
 storageConnectionString=$(echo $stateJson | jq -r '.state.storage.connectionString')
 storageContainerName=$(echo $stateJson | jq -r '.vars.storage.containerName')
 testCloudProvider=$(echo $metaJson | jq -r '.meta.cloud_provider')
@@ -39,10 +39,9 @@ statsPaths=(
   "vpn"
 )
 metaPath="meta"
-metaFilePatterns=(
-  "run.meta.json"
-  "PubSub.docker-compose.deployed.yml"
-)
+metaFilePattern="run.meta.json"
+dockerComposePath="docker"
+dockerComposeFilePattern="PubSub.docker-compose.deployed.yml"
 
 ##############################################################################################################################
 # Upload stats
@@ -62,16 +61,25 @@ done
 # Upload meta
 echo " >>> Upload Tests Results Meta..."
 destPath="$prefixPath/$metaPath"
-for metaFilePattern in ${metaFilePatterns[@]}; do
-  resp=$(az storage blob upload-batch \
-    --connection-string $storageConnectionString \
-    --destination $storageContainerName \
-    --source $runDir \
-    --pattern "$metaFilePattern" \
-    --destination-path "$destPath" \
-    --verbose)
-  echo " >>> Success."
-done
+resp=$(az storage blob upload-batch \
+  --connection-string $storageConnectionString \
+  --destination $storageContainerName \
+  --source $runDir \
+  --pattern "$metaFilePattern" \
+  --destination-path "$destPath" \
+  --verbose)
+echo " >>> Success."
+
+echo " >>> Upload Tests Docker Compose Template ..."
+destPath="$prefixPath/$dockerComposePath"
+resp=$(az storage blob upload-batch \
+  --connection-string $storageConnectionString \
+  --destination $storageContainerName \
+  --source $runDir \
+  --pattern "$dockerComposeFilePattern" \
+  --destination-path "$destPath" \
+  --verbose)
+echo " >>> Success."
 
 ##############################################################################################################################
 # Manual step
@@ -88,10 +96,12 @@ echo "#   - go to: $dataIngestionUri"
 echo "#   - Ping: table: ping, source type: From container"
 echo "#   - Latency: table: latency, source type: From container"
 echo "#   - Vpn: table: vpn, source type: From container"
+echo "#   - Meta: table: meta, source type: From container"
 echo "#   - Links to storage: "
 for statsPath in ${statsPaths[@]}; do
   echo "    $statsPath: $link2StorageBase/$statsPath?$sasToken"
 done
+  echo "    $metaPath: $link2StorageBase/$metaPath?$sasToken"
 echo
 
 ###